refactor(home): align imports and subpath barrel with Classic pattern
Address Copilot review on PR #1494:

- src/home.ts: expose `HomeAPIAdapter as APIAdapter` for parity with
  the Classic subpath barrel (consumers using `melcloud-api/home`
  can now reach `home.APIAdapter` for mocking).
- src/facades/{home-manager,home-device-ata}.ts: import
  `HomeAPIAdapter` from `../api/index.ts` (the API module's public
  barrel) instead of the deep `../api/home-types.ts` path. Matches
  what Classic facades already do for `ClassicAPIAdapter` and keeps
  facade code dependent on public API surfaces only.
